add fix tests for task deletion

+++ task_manager/tasks/views.py
@@ -40,7 +40,7 @@ class DeleteView(LoginRequiredMixin, View):
     def get(self, request, *args, **kwargs):
         task_id = kwargs.get("pk")
         task = get_object_or_404(Task, id=task_id)
-        if task.author is not request.user:
+        if not (task.author == request.user):
             messages.add_message(request, messages.INFO, _('A task can only be deleted by its author'))
             return redirect('tasks.index')
 
@@ -51,6 +51,10 @@ class DeleteView(LoginRequiredMixin, View):
     def post(self, request, *args, **kwargs):
         task_id = kwargs.get("pk")
         task = get_object_or_404(Task, id=task_id)
+
+        if not (task.author == request.user):
+            messages.add_message(request, messages.INFO, _('A task can only be deleted by its author'))
+            return redirect('tasks.index')
 
         task.delete()
         messages.add_message(request, messages.INFO, _("Task successfully deleted"))
